CAD_models/requirement.txt
fastapi
uvicorn
torch
numpy
scipy
wfdb
pandas
matplotlib
python-multipart
neurokit2
pymongo
motor
python-jose[cryptography]
email-validator
authlib
python-dotenv
google-auth
pyjwt
opencv-python
bcrypt==4.0.1
passlib[bcrypt]==1.7.4